refactor(StyledCollectiveCard): Centralize memberships and discover
